app/requirements.txt
beautifulsoup4==4.9.0
bs4==0.0.1
Click==7.0
numpy==1.18.1
opencv-python==4.2.0.34
pandas==1.0.3
Pillow==5.1.0
selenium==3.141.0
typing-extensions==3.7.4.1
